refactor(i18n): promote core/localization/ → i18n/ (top-level)
MessageCatalog, L, LV, IntlFormatter are public API — their home
should reflect that. Update 18 import references across _runtime,
dialogs, widgets, core capabilities, and bootstack/__init__.py.
Remove localization from core/__init__.py.
