+++ src/bootstack/i18n/specs.py
