+++ src/bootstack/widgets/_parts/spinnerentry_part.py
@@ -8,7 +8,7 @@ both text and numeric values.
 from tkinter import Event, TclError
 from typing import Any, Callable
 
-from bootstack.core.localization import MessageCatalog, IntlFormatter
+from bootstack.i18n import MessageCatalog, IntlFormatter
 from bootstack.widgets.primitives.spinbox import Spinbox
 from bootstack.widgets.mixins import ValidationMixin
 from bootstack.widgets.mixins.configure_mixin import configure_delegate
